+++ src/DDTrace/Integrations/Eloquent/EloquentIntegration.php
@@ -3,8 +3,8 @@
 namespace DDTrace\Integrations\Eloquent;
 
 use DDTrace\Integrations\Integration;
-use DDTrace\Tags;
-use DDTrace\Types;
+use DDTrace\Tag;
+use DDTrace\Type;
 use DDTrace\Util\TryCatchFinally;
 use OpenTracing\GlobalTracer;
 
@@ -25,9 +25,9 @@ class EloquentIntegration
             $scope = GlobalTracer::get()->startActiveSpan('eloquent.get');
             $span = $scope->getSpan();
             $sql = $this->getQuery()->toSql();
-            $span->setTag(Tags\RESOURCE_NAME, $sql);
-            $span->setTag(Tags\DB_STATEMENT, $sql);
-            $span->setTag(Tags\SPAN_TYPE, Types\SQL);
+            $span->setTag(Tag::RESOURCE_NAME, $sql);
+            $span->setTag(Tag::DB_STATEMENT, $sql);
+            $span->setTag(Tag::SPAN_TYPE, Type::SQL);
 
             return TryCatchFinally::executePublicMethod($scope, $this, 'getModels', $args);
         });
@@ -39,9 +39,9 @@ class EloquentIntegration
             $scope = GlobalTracer::get()->startActiveSpan('eloquent.insert');
             $span = $scope->getSpan();
             $sql = $eloquentQueryBuilder->getQuery()->toSql();
-            $span->setTag(Tags\RESOURCE_NAME, $sql);
-            $span->setTag(Tags\DB_STATEMENT, $sql);
-            $span->setTag(Tags\SPAN_TYPE, Types\SQL);
+            $span->setTag(Tag::RESOURCE_NAME, $sql);
+            $span->setTag(Tag::DB_STATEMENT, $sql);
+            $span->setTag(Tag::SPAN_TYPE, Type::SQL);
 
             return TryCatchFinally::executePublicMethod($scope, $this, 'performInsert', $args);
         });
@@ -53,9 +53,9 @@ class EloquentIntegration
             $scope = GlobalTracer::get()->startActiveSpan('eloquent.update');
             $span = $scope->getSpan();
             $sql = $eloquentQueryBuilder->getQuery()->toSql();
-            $span->setTag(Tags\RESOURCE_NAME, $sql);
-            $span->setTag(Tags\DB_STATEMENT, $sql);
-            $span->setTag(Tags\SPAN_TYPE, Types\SQL);
+            $span->setTag(Tag::RESOURCE_NAME, $sql);
+            $span->setTag(Tag::DB_STATEMENT, $sql);
+            $span->setTag(Tag::SPAN_TYPE, Type::SQL);
 
             return TryCatchFinally::executePublicMethod($scope, $this, 'performUpdate', $args);
         });
@@ -63,7 +63,7 @@ class EloquentIntegration
         // public function delete()
         dd_trace('Illuminate\Database\Eloquent\Model', 'delete', function () {
             $scope = GlobalTracer::get()->startActiveSpan('eloquent.delete');
-            $scope->getSpan()->setTag(Tags\SPAN_TYPE, Types\SQL);
+            $scope->getSpan()->setTag(Tag::SPAN_TYPE, Type::SQL);
 
             return TryCatchFinally::executePublicMethod($scope, $this, 'delete', []);
         });