+++ src/DDTrace/Integrations/Mysqli/MysqliIntegration.php
@@ -3,8 +3,8 @@
 namespace DDTrace\Integrations\Mysqli;
 
 use DDTrace\Integrations\Integration;
-use DDTrace\Tags;
-use DDTrace\Types;
+use DDTrace\Tag;
+use DDTrace\Type;
 use DDTrace\Util\ObjectKVStore;
 use DDTrace\Util\TryCatchFinally;
 use OpenTracing\GlobalTracer;
@@ -354,9 +354,9 @@ class MysqliIntegration
         $scope = GlobalTracer::get()->startActiveSpan($operationName);
         /** @var \DDTrace\Span $span */
         $span = $scope->getSpan();
-        $span->setTag(Tags\SPAN_TYPE, Types\SQL);
-        $span->setTag(Tags\SERVICE_NAME, 'mysqli');
-        $span->setTag(Tags\RESOURCE_NAME, $resource);
+        $span->setTag(Tag::SPAN_TYPE, Type::SQL);
+        $span->setTag(Tag::SERVICE_NAME, 'mysqli');
+        $span->setTag(Tag::RESOURCE_NAME, $resource);
         return $scope;
     }