+++ src/DDTrace/Integrations/Predis/PredisIntegration.php
@@ -3,8 +3,8 @@
 namespace DDTrace\Integrations\Predis;
 
 use DDTrace\Integrations\Integration;
-use DDTrace\Tags;
-use DDTrace\Types;
+use DDTrace\Tag;
+use DDTrace\Type;
 use DDTrace\Util\TryCatchFinally;
 use OpenTracing\GlobalTracer;
 use Predis\Configuration\OptionsInterface;
@@ -38,9 +38,9 @@ class PredisIntegration
             $args = func_get_args();
             $scope = GlobalTracer::get()->startActiveSpan('Predis.Client.__construct');
             $span = $scope->getSpan();
-            $span->setTag(Tags\SPAN_TYPE, Types\CACHE);
-            $span->setTag(Tags\SERVICE_NAME, 'redis');
-            $span->setTag(Tags\RESOURCE_NAME, 'Predis.Client.__construct');
+            $span->setTag(Tag::SPAN_TYPE, Type::CACHE);
+            $span->setTag(Tag::SERVICE_NAME, 'redis');
+            $span->setTag(Tag::RESOURCE_NAME, 'Predis.Client.__construct');
 
             $thrown = null;
             try {
@@ -64,9 +64,9 @@ class PredisIntegration
         dd_trace('\Predis\Client', 'connect', function () {
             $scope = GlobalTracer::get()->startActiveSpan('Predis.Client.connect');
             $span = $scope->getSpan();
-            $span->setTag(Tags\SPAN_TYPE, Types\CACHE);
-            $span->setTag(Tags\SERVICE_NAME, 'redis');
-            $span->setTag(Tags\RESOURCE_NAME, 'Predis.Client.connect');
+            $span->setTag(Tag::SPAN_TYPE, Type::CACHE);
+            $span->setTag(Tag::SERVICE_NAME, 'redis');
+            $span->setTag(Tag::RESOURCE_NAME, 'Predis.Client.connect');
             PredisIntegration::setConnectionTags($this, $span);
 
             return TryCatchFinally::executePublicMethod($scope, $this, 'connect', []);
@@ -80,11 +80,11 @@ class PredisIntegration
 
             $scope = GlobalTracer::get()->startActiveSpan('Predis.Client.executeCommand');
             $span = $scope->getSpan();
-            $span->setTag(Tags\SPAN_TYPE, Types\CACHE);
-            $span->setTag(Tags\SERVICE_NAME, 'redis');
+            $span->setTag(Tag::SPAN_TYPE, Type::CACHE);
+            $span->setTag(Tag::SERVICE_NAME, 'redis');
             $span->setTag('redis.raw_command', $query);
             $span->setTag('redis.args_length', count($arguments));
-            $span->setTag(Tags\RESOURCE_NAME, $query);
+            $span->setTag(Tag::RESOURCE_NAME, $query);
             PredisIntegration::setConnectionTags($this, $span);
 
             return TryCatchFinally::executePublicMethod($scope, $this, 'executeCommand', [$command]);
@@ -98,11 +98,11 @@ class PredisIntegration
 
                 $scope = GlobalTracer::get()->startActiveSpan('Predis.Client.executeRaw');
                 $span = $scope->getSpan();
-                $span->setTag(Tags\SPAN_TYPE, Types\CACHE);
-                $span->setTag(Tags\SERVICE_NAME, 'redis');
+                $span->setTag(Tag::SPAN_TYPE, Type::CACHE);
+                $span->setTag(Tag::SERVICE_NAME, 'redis');
                 $span->setTag('redis.raw_command', $query);
                 $span->setTag('redis.args_length', count($arguments));
-                $span->setTag(Tags\RESOURCE_NAME, $query);
+                $span->setTag(Tag::RESOURCE_NAME, $query);
                 PredisIntegration::setConnectionTags($this, $span);
 
                 // PHP 5.4 compatible try-catch-finally block.
@@ -132,8 +132,8 @@ class PredisIntegration
             dd_trace('\Predis\Pipeline\Pipeline', 'executePipeline', function ($connection, $commands) {
                 $scope = GlobalTracer::get()->startActiveSpan('Predis.Pipeline.executePipeline');
                 $span = $scope->getSpan();
-                $span->setTag(Tags\SPAN_TYPE, Types\CACHE);
-                $span->setTag(Tags\SERVICE_NAME, 'redis');
+                $span->setTag(Tag::SPAN_TYPE, Type::CACHE);
+                $span->setTag(Tag::SERVICE_NAME, 'redis');
                 $span->setTag('redis.pipeline_length', count($commands));
                 PredisIntegration::setConnectionTags($this, $span);
 
@@ -169,8 +169,8 @@ class PredisIntegration
         try {
             $identifier = (string)$predis->getConnection();
             list($host, $port) = explode(':', $identifier);
-            $tags[Tags\TARGET_HOST] = $host;
-            $tags[Tags\TARGET_PORT] = $port;
+            $tags[Tag::TARGET_HOST] = $host;
+            $tags[Tag::TARGET_PORT] = $port;
         } catch (\Exception $e) {
         }