+++ tests/Unit/SpanTest.php
@@ -4,7 +4,7 @@ namespace DDTrace\Tests\Unit;
 
 use DDTrace\Span;
 use DDTrace\SpanContext;
-use DDTrace\Tags;
+use DDTrace\Tag;
 use Exception;
 use PHPUnit\Framework;
 
@@ -52,21 +52,21 @@ final class SpanTest extends Framework\TestCase
     public function testSpanTagWithErrorCreatesExpectedTags()
     {
         $span = $this->createSpan();
-        $span->setTag(Tags\ERROR, new Exception(self::EXCEPTION_MESSAGE));
+        $span->setTag(Tag::ERROR, new Exception(self::EXCEPTION_MESSAGE));
 
         $this->assertTrue($span->hasError());
-        $this->assertEquals($span->getTag(Tags\ERROR_MSG), self::EXCEPTION_MESSAGE);
-        $this->assertEquals($span->getTag(Tags\ERROR_TYPE), 'Exception');
+        $this->assertEquals($span->getTag(Tag::ERROR_MSG), self::EXCEPTION_MESSAGE);
+        $this->assertEquals($span->getTag(Tag::ERROR_TYPE), 'Exception');
     }
 
     public function testSpanTagWithErrorBoolProperlyMarksError()
     {
         $span = $this->createSpan();
 
-        $span->setTag(Tags\ERROR, true);
+        $span->setTag(Tag::ERROR, true);
         $this->assertTrue($span->hasError());
 
-        $span->setTag(Tags\ERROR, false);
+        $span->setTag(Tag::ERROR, false);
         $this->assertFalse($span->hasError());
     }
 
@@ -74,10 +74,10 @@ final class SpanTest extends Framework\TestCase
     {
         $span = $this->createSpan();
 
-        $span->log([Tags\LOG_ERROR => true]);
+        $span->log([Tag::LOG_ERROR => true]);
         $this->assertTrue($span->hasError());
 
-        $span->log([Tags\LOG_ERROR => false]);
+        $span->log([Tag::LOG_ERROR => false]);
         $this->assertFalse($span->hasError());
     }
 
@@ -85,7 +85,7 @@ final class SpanTest extends Framework\TestCase
     {
         $span = $this->createSpan();
 
-        $span->log([Tags\LOG_EVENT => 'error']);
+        $span->log([Tag::LOG_EVENT => 'error']);
         $this->assertTrue($span->hasError());
     }
 
@@ -93,42 +93,42 @@ final class SpanTest extends Framework\TestCase
     {
         $span = $this->createSpan();
 
-        $span->log([Tags\LOG_EVENT => 'some other event']);
+        $span->log([Tag::LOG_EVENT => 'some other event']);
         $this->assertFalse($span->hasError());
 
-        $span->log([Tags\LOG_ERROR => false]);
+        $span->log([Tag::LOG_ERROR => false]);
         $this->assertFalse($span->hasError());
     }
 
 
     public function testSpanLogWithErrorCreatesExpectedTags()
     {
-        foreach ([Tags\LOG_ERROR, Tags\LOG_ERROR_OBJECT] as $key) {
+        foreach ([Tag::LOG_ERROR, Tag::LOG_ERROR_OBJECT] as $key) {
             $span = $this->createSpan();
             $span->log([$key => new Exception(self::EXCEPTION_MESSAGE)]);
 
             $this->assertTrue($span->hasError());
-            $this->assertEquals($span->getTag(Tags\ERROR_MSG), self::EXCEPTION_MESSAGE);
-            $this->assertEquals($span->getTag(Tags\ERROR_TYPE), 'Exception');
+            $this->assertEquals($span->getTag(Tag::ERROR_MSG), self::EXCEPTION_MESSAGE);
+            $this->assertEquals($span->getTag(Tag::ERROR_TYPE), 'Exception');
         }
     }
 
     public function testSpanLogStackAddsExpectedTag()
     {
         $span = $this->createSpan();
-        $span->log([Tags\LOG_STACK => self::DUMMY_STACK_TRACE]);
+        $span->log([Tag::LOG_STACK => self::DUMMY_STACK_TRACE]);
 
         $this->assertFalse($span->hasError());
-        $this->assertEquals($span->getTag(Tags\ERROR_STACK), self::DUMMY_STACK_TRACE);
+        $this->assertEquals($span->getTag(Tag::ERROR_STACK), self::DUMMY_STACK_TRACE);
     }
 
     public function testSpanLogMessageAddsExpectedTag()
     {
         $span = $this->createSpan();
-        $span->log([Tags\LOG_MESSAGE => self::EXCEPTION_MESSAGE]);
+        $span->log([Tag::LOG_MESSAGE => self::EXCEPTION_MESSAGE]);
 
         $this->assertFalse($span->hasError());
-        $this->assertEquals($span->getTag(Tags\ERROR_MSG), self::EXCEPTION_MESSAGE);
+        $this->assertEquals($span->getTag(Tag::ERROR_MSG), self::EXCEPTION_MESSAGE);
     }
 
     public function testSpanErrorAddsExpectedTags()
@@ -137,9 +137,9 @@ final class SpanTest extends Framework\TestCase
         $span->setError(new Exception(self::EXCEPTION_MESSAGE));
 
         $this->assertTrue($span->hasError());
-        $this->assertEquals($span->getTag(Tags\ERROR_MSG), self::EXCEPTION_MESSAGE);
-        $this->assertNotEmpty($span->getTag(Tags\ERROR_STACK));
-        $this->assertEquals($span->getTag(Tags\ERROR_TYPE), 'Exception');
+        $this->assertEquals($span->getTag(Tag::ERROR_MSG), self::EXCEPTION_MESSAGE);
+        $this->assertNotEmpty($span->getTag(Tag::ERROR_STACK));
+        $this->assertEquals($span->getTag(Tag::ERROR_TYPE), 'Exception');
     }
 
     public function testSpanErrorRemainsImmutableAfterFinishing()
@@ -164,9 +164,9 @@ final class SpanTest extends Framework\TestCase
     public function testAddCustomTagsSuccess()
     {
         $span = $this->createSpan();
-        $span->setTag(Tags\SERVICE_NAME, self::ANOTHER_SERVICE);
-        $span->setTag(Tags\RESOURCE_NAME, self::ANOTHER_RESOURCE);
-        $span->setTag(Tags\SPAN_TYPE, self::ANOTHER_TYPE);
+        $span->setTag(Tag::SERVICE_NAME, self::ANOTHER_SERVICE);
+        $span->setTag(Tag::RESOURCE_NAME, self::ANOTHER_RESOURCE);
+        $span->setTag(Tag::SPAN_TYPE, self::ANOTHER_TYPE);
 
         $this->assertEquals(self::ANOTHER_SERVICE, $span->getService());
         $this->assertEquals(self::ANOTHER_RESOURCE, $span->getResource());